implement pagerduty module
This uses the go-pagerduty library and the oncall api. The response from
the oncall api is not ideally structured, which is why this client
implementation restructures it.

This addition still needs:
to not get the API key from env vars
to be able to filter by Users and Escalation Policies (passed to the api
request)